👷 (docker): Add Dockerfile for Node app
A new Dockerfile has been included to support the deployment of a Node application. This Dockerfile starts from a base image of Node v20, installs the required dependencies, builds the application, and sets it up for running in a Docker container. It makes use of the pnpm package manager for efficient dependency installation.