♻️ (home): Move nav-main.svelte contents to +page.svelte
The nav-main.svelte file has been deleted and its contents have been transferred to +page.svelte. This move is done to simplify the codebase and avoid unnecessary import statements. The change will not affect the functionality of the navigation system and the social media links, as they are simply relocated.