✨ (disco): New slub and new utility class
The date formatting function has been moved to a newly created utility class, `FormatUtils`. This is done to reduce the redundancy of the function, improve the organization of the code, and increase its maintainability. This class has been applied to multiple routes in the application (music page and slug). The code reference for the background image was also updated for better sourcing.