✨ (logo) Added logo component and replaced previous text on svelte page.
This commit introduces a new logo component and replaces the previous text "Opus Iræ" on the svelte page with this logo. The logo is imported and applied on the header of the page. Other styles and spacings on the page are also adjusted accordingly for the proper placement and view of the logo. The logo component itself consists of an SVG enclosed in a template.